Rename timeseriestools -> timeseries
