+++ climakitae/__init__.py
@@ -3,7 +3,7 @@ from climakitae.ui.select import Select
 from climakitae.core.data_load import load
 from climakitae.core.data_view import view
 from climakitae.ui.export import Export
-from climakitae.tools.timeseriestools import TimeSeries
+from climakitae.tools.timeseries import TimeSeries
 from climakitae.explore import Explore
 
 try:
