+++ tests/timeseriestools/test_timeseriestools_monthly.py
@@ -5,7 +5,7 @@ run through `transform_data` without error, and that the data has been transform
 (not equal to original values), but does not test for exact expected values.
 """
 
-from climakitae.tools import timeseriestools as tst
+from climakitae.tools import timeseries as tst
 import datetime as dt
 import pytest
 import os
